requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
blis==1.3.0
build==1.3.0
CacheControl==0.14.4
cachetools==6.2.1
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.5
cleo==2.1.0
click==8.2.1
click-default-group==1.2.4
cloudpathlib==0.22.0
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
confection==0.1.5
crashtest==0.4.1
cryptography==46.0.3
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
dulwich==0.24.10
email-validator==2.3.0
emoji==2.15.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.3.1
executing==2.2.1
faiss-cpu==1.13.0
fastapi==0.114.0
filelock==3.20.0
findpython==0.7.1
Flask==2.3.3
flask-cors==6.0.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
greenlet==3.2.4
groq==0.37.0
grpcio==1.76.0
grpcio-status==1.76.0
gymnasium==1.2.2
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
installer==0.7.0
ipykernel==7.1.0
ipython==8.12.3
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.16
keyring==25.7.0
kubernetes==34.1.0
langchain==1.2.3
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langcodes==3.5.0
langdetect==1.0.9
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.2
language_data==1.3.0
lark==1.3.1
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
mmh3==5.2.0
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==2.11.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
notebook_shim==0.2.4
npm==0.1.1
numpy==1.26.4
oauthlib==3.3.1
olefile==0.47
ollama==0.6.0
onnxruntime==1.23.2
openai==1.109.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging==23.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pbs-installer==2025.10.31
pdfminer.six==20251107
pdfplumber==0.11.8
pettingzoo==1.25.0
pillow==12.0.0
pkginfo==1.12.1.2
platformdirs==4.5.0
pluggy==1.6.0
poetry==2.2.1
poetry-core==2.2.1
postgrest==2.27.0
posthog==5.4.0
preshed==3.0.10
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
pyairtable==3.3.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==6.4.0
pypdfium2==5.1.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyroaring==1.0.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-iso639==2025.11.16
python-json-logger==4.0.0
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
pywin32-ctypes==0.2.3
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
realtime==2.27.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
Send2Trash==1.8.3
sentence-transformers==5.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
sqlite-utils==3.38
srsly==2.5.1
stack-data==0.6.3
starlette==0.38.6
storage3==2.27.0
streamlit==1.51.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.20
unstructured-client==0.42.4
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.3.0
uuid6==2024.7.10
uuid_utils==0.12.0
uv==0.9.10
uvicorn==0.23.1
virtualenv==20.35.4
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
weasel==0.4.1
webcolors==25.10.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
